Release 0.8.1 (#27)
* MNT: Re-rendered with conda-build 3.21.4, conda-smithy 3.9.0, and conda-forge-pinning 2021.03.02.20.03.28

* Factor out package URL location

This helps simplify updating if this changes between versions.

* Release 0.8.1

* Package the license file

As this is a whl that includes the license file within it and installs
it alongside the package, just grab the license file from where it gets
installed. This way we can ensure it is included in the Conda package
along with other package licenses.

* Use `__unix` metapackage to make this `noarch`

As this doesn't work on Windows, but otherwise is a `noarch` package,
require the `__unix` metapackage in `run`. This ensures the package
cannot be installed on OSes that are not Unix.

* MNT: Re-rendered with conda-build 3.21.4, conda-smithy 3.9.0, and conda-forge-pinning 2021.03.02.20.03.28

Now that this is a `noarch` package, re-render to update the feedstock
accordingly.

* Revert "Use `__unix` metapackage to make this `noarch`"

This reverts commit fa02704809d81956871b9eee58fbe1001aea5118.

* MNT: Re-rendered with conda-build 3.21.4, conda-smithy 3.9.0, and conda-forge-pinning 2021.03.02.20.03.28

Re-render now the `noarch` is disabled to update the feedstock
accordingly.